Switch to using rtmbot package
Most of the work here was work needed to conform to
rtmbot's expectations.

It expects the plugins package to be in the directory in which
rtmbot is run, so the plugins folder is moved.

rtmbot doesn't print lots of messages to the log, so remove
LOG_LEVEL parameter from the template.

Add new ACTIVE_PLUGINS parameter to enable the SystemD plugin.

Reorganize plugins/systemd.py so that the process_message function
belongs to a class that extends rtmbot.core.Plugin.

Update Makefile to run rtmbot.