Don't automatically install dependencies
Do not do a pip -r requirements.txt whenever it changes, since
that would require the application to run as another user (since
it doesn't have access to its own code).

Use a file inside venv to ensure that we don't redo targets just
because the directory is treated as newer than something else.